requirements.txt
langchain==0.3.0
langgraph==0.2.49
langchain-groq==0.2.0
streamlit==1.56.0
python-dotenv==1.0.0
ddgs==9.0.0
requests==2.32.3
geopy==2.4.0
pydantic==2.9.0
httpx==0.27.0